Added build option to treat exit != 0 as error
This is a compromize of the previous solution with past behavior.  The
key issue is that, when a build of Go code fails, the error is not
immediately obvious in the console output.  This adds an
"expectExitZero" option to build actions which, when set to true, causes
LiteBuild to treat a non-zero exit as an error.  It is currently only
enabled for "go build" and "go install" actions.  Consequently, failed
builds appear in red.