Auto-complete ignores strings and comments
This was implemented using the spell-checking system already in place in
the Kate files.  The idea is that if a region has spell-check enabled
(e.g. comments and strings) then it is written in a natural language,
and thus it is not code.

This also has the positive side effect of not including words found in
strings in regions that ARE auto-completed.